API module refactoring
