+++ src/api/api/article/articleValidate.js
@@ -1,11 +1,11 @@
-const {HttpCode} = require('../../../../../constants');
+const {HttpCode} = require('../../../constants');
 
 const articleKeys = ["title", "announce", "fullText", "category", "createdDate", "comments"]
 
 module.exports = (req, res, next) => {
   const newOffer = req.body;
   const keys = Object.keys(newOffer);
-  const keysExists = keys.every((key) => articleKeys.includes(key));
+  const keysExists = articleKeys.every((key) => keys.includes(key));
 
   if (!keysExists) {
     return res.status(HttpCode.BAD_REQUEST)
