+++ src/api/lib/getMockData.js
@@ -1,5 +1,6 @@
+const chalk = require(`chalk`);
 const fs = require(`fs`).promises;
-const FILENAME = __dirname + '/../cli/generate/mock.json';
+const MOCK_FILE = __dirname +`/../../service/cli/generate/mock.json`;
 let data = null;
 
 const getMockData = async () => {
@@ -8,10 +9,10 @@ const getMockData = async () => {
   }
 
   try {
-    const fileContent = await fs.readFile(FILENAME);
+    const fileContent = await fs.readFile(MOCK_FILE);
     data = JSON.parse(fileContent);
   } catch (err) {
-    console.log(err);
+    console.log(chalk.red(err));
     return Promise.reject(err);
   }
