+++ src/service/cli/generate/data/categories.txt
