+++ src/service/cli/generate/data/sentences.txt
