+++ src/service/cli/help/help.txt
@@ -1,7 +1,4 @@
-Программа запускает http-сервер и формирует файл с данными для API.
-
-    Гайд:
-    server <command>
+Программа формирует файл с данными для API.
 
     Команды:
     --version:            выводит номер версии
