+++ src/service/cli/index.js
@@ -3,13 +3,11 @@
 const version = require(`./version`);
 const help = require(`./help`);
 const generate = require(`./generate`);
-const server = require(`./server`);
 
 const cli = {
   [version.name]: version,
   [help.name]: help,
   [generate.name]: generate,
-  [server.name]: server,
 };
 
 module.exports = {
