--- src/service/readme.md
@@ -1,1 +1,0 @@
-# Папка для сценариев API-сервиса
